tests: be safer when using QDir::removeRecursively() on teardown
if initTestCase fails before setting any of them, they will be left
with the default value, which is the current directory (i.e. the build
directory for autotests) -- not a good idea to nuke part of the build
directory like that :)

+++ autotests/kiconloader_unittest.cpp
@@ -146,8 +146,12 @@ private Q_SLOTS:
 
     void cleanupTestCase()
     {
-        testIconsDir.removeRecursively();
-        appDataDir.removeRecursively();
+        if (testIconsDir != QDir()) {
+            testIconsDir.removeRecursively();
+        }
+        if (appDataDir != QDir()) {
+            appDataDir.removeRecursively();
+        }
     }
 
     void init()
